Fix up Discordlink system, SQL cleanup (#860)
Fixes the discordlink system and cleans up queries to be more performant (and actually correct)

Changes the discordlink model, queries and result -> model associations to be inline with the DB schema instead of nonsense

Changes panic bunker living minutes defaults to be -1, which means nothing gets blocked. This is currently a stopgap to make the discord module of the PB work and should be properly fixed at a later point.