New mob fire! (#66746)
This is a port of TauCeti's pending fire icons, I really like them, and have been given spritetainer approval to port these.

This PR also changes how fire is handled, or at least, returns the way it used to be handled, while also updating it to use species to choose what kind of fire overlay to use.